PR : Seperated concerns, and simplified Art
Discussed and concluded that we want to sacrifice some fidelity to make the sample more extensible, so removing extra art and simplifying the parts that are there.

Also addressed all points on the PR, Seperated the PartyInfoView into PartyList and PartyView.
Revised how we subscribe and unsubscribe to lobby events.
Added Error Panel
Overall Cleanup.